requirements.txt
openai>=1.3.0
chromadb>=0.4.0
sentence-transformers
tiktoken
fastapi
uvicorn
streamlit
python-dotenv
pypdf
requests
numpy
httpx
pytest
scikit-learn
